[MDETR] Separate losses from model
Summary:
This PR takes the losses and matcher out of MDETRForVQA and MDETRForPhraseGrounding classes.
We keep contrastive projections in both, even though they are not technically necessary for eval.
The alternative is to add a flag but that introduces a lot of copypasta. Also modified the eval
and fine-tuning loops and added a builder for the loss class.

Test plan:
Phrase grounding eval:

```
CUBLAS_WORKSPACE_CONFIG=:4096:8 python -m torch.distributed.launch --nproc_per_node=2 --use_env phrase_grounding.py
--resume https://pytorch.s3.amazonaws.com/models/multimodal/mdetr/pretrained_resnet101_checkpoint.pth --ema --eval
--dataset_config phrase_grounding.json
```
Gives recall@1 82.4% as expected

VQA eval:

```
CUBLAS_WORKSPACE_CONFIG=:4096:8 python -m torch.distributed.launch --nproc_per_node=4 --use_env vqa_eval.py
--resume https://pytorch.s3.amazonaws.com/models/multimodal/mdetr/gqa_resnet101_checkpoint.pth --ema --eval
--dataset_config vqa.json
```
Gives total accuracy 61.6% as expected

VQA fine-tuning:
```
CUBLAS_WORKSPACE_CONFIG=:4096:8 python -m torch.distributed.launch --nproc_per_node=4 --use_env vqa_finetune.py
--dataset_config vqa.json --ema --epochs 125 --epoch_chunks 3 --lr_drop 150
--load /data/home/ebs/data/mdetr/checkpoints/pretrained_resnet101_checkpoint.pth --qa_loss_coef 25 --batch_size 4
--lr 1.4e-4 --lr_backbone 1.4e-5 --text_encoder_lr 7e-5
```
Logged losses from first ten iterations:
```
0 tensor(199.3950, device='cuda:0', grad_fn=<AddBackward0>)
1 tensor(178.6834, device='cuda:0', grad_fn=<AddBackward0>)
2 tensor(187.6183, device='cuda:0', grad_fn=<AddBackward0>)
3 tensor(164.5415, device='cuda:0', grad_fn=<AddBackward0>)
4 tensor(352.6355, device='cuda:0', grad_fn=<AddBackward0>)
5 tensor(220.8754, device='cuda:0', grad_fn=<AddBackward0>)
6 tensor(178.6945, device='cuda:0', grad_fn=<AddBackward0>)
7 tensor(342.7989, device='cuda:0', grad_fn=<AddBackward0>)
8 tensor(169.4948, device='cuda:0', grad_fn=<AddBackward0>)
9 tensor(186.1401, device='cuda:0', grad_fn=<AddBackward0>)
10 tensor(126.2118, device='cuda:0', grad_fn=<AddBackward0>),
```
which matches the results on main branch.

[ghstack-poisoned]
