+++ torchmultimodal/models/mdetr/model.py
@@ -4,10 +4,13 @@
 # This source code is licensed under the BSD-style license found in the
 # LICENSE file in the root directory of this source tree.
 
+
 import math
-from typing import List, NamedTuple, Optional, Tuple
+from collections import OrderedDict
+from typing import Dict, List, NamedTuple, Optional, Tuple
 
 import torch
+import torch.nn.functional as F
 from torch import nn, Tensor
 from torchmultimodal.models.mdetr.image_encoder import (
     mdetr_resnet101_backbone,
@@ -257,3 +260,236 @@ def mdetr_resnet101(
         extra_query_embeddings,
     )
     return mdetr
+
+
+class MultiHead(nn.Module):
+    def __init__(self, heads: nn.ModuleDict):
+        super().__init__()
+        self.heads = heads
+
+    def forward(
+        self,
+        embeddings: Tensor,
+    ) -> Dict[str, Tensor]:
+        if embeddings.size(0) != len(self.heads):
+            raise ValueError("Number of embeddings must equal number of heads")
+        out = OrderedDict()
+        for (head_name, head), embedding in zip(self.heads.items(), embeddings):
+            out[head_name] = head(embedding)
+
+        return out
+
+
+def mdetr_gqa_heads(hidden_dim: int = 256) -> MultiHead:
+    answer_type_head = nn.Linear(hidden_dim, 5)  # Number of answer types
+    answer_obj_head = nn.Linear(hidden_dim, 3)
+    answer_attr_head = nn.Linear(hidden_dim, 403)
+    answer_rel_head = nn.Linear(hidden_dim, 1594)
+    answer_global_head = nn.Linear(hidden_dim, 111)
+    answer_cat_head = nn.Linear(hidden_dim, 678)
+    heads = nn.ModuleDict(
+        {
+            "answer_type": answer_type_head,
+            "answer_obj": answer_obj_head,
+            "answer_rel": answer_rel_head,
+            "answer_attr": answer_attr_head,
+            "answer_cat": answer_cat_head,
+            "answer_global": answer_global_head,
+        }
+    )
+    return MultiHead(heads)
+
+
+class ContrastiveEmbeddingsOutput(NamedTuple):
+    query_embeddings: Tensor
+    token_embeddings: Tensor
+
+
+class MDETRVQAOutput(NamedTuple):
+    model_output: MDETRModelOutput
+    vqa_preds: Dict[str, Tensor]
+    contrastive_embeddings: ContrastiveEmbeddingsOutput
+
+
+class MDETRForVQA(nn.Module):
+    def __init__(
+        self,
+        model: MDETR,
+        vqa_heads: MultiHead,
+        contrastive_alignment_image_projection: nn.Module,
+        contrastive_alignment_text_projection: nn.Module,
+    ):
+        super().__init__()
+        self.model = model
+        self.vqa_heads = vqa_heads
+        if self.model.extra_query_embeddings is None:
+            raise ValueError("MDETRForVQA requires extra query embeddings ")
+        if self.model.extra_query_embeddings.num_embeddings != len(
+            self.vqa_heads.heads.keys()
+        ):
+            raise ValueError("Number of heads must match number of QA embeddings")
+
+        self.contrastive_alignment_image_projection = (
+            contrastive_alignment_image_projection
+        )
+        self.contrastive_alignment_text_projection = (
+            contrastive_alignment_text_projection
+        )
+
+    def forward(
+        self,
+        images: List[Tensor],
+        text: List[Tensor],
+    ) -> MDETRVQAOutput:
+        # Calculate MDETR model outputs
+        model_output = self.model(images, text)
+
+        final_hidden_state = model_output.transformer_output.decoder_hidden_states[-1]
+
+        # Perform projections for contrastive alignment loss.
+        contrastive_query_embeddings = F.normalize(
+            self.contrastive_alignment_image_projection(final_hidden_state),
+            p=2,
+            dim=-1,
+        )
+        contrastive_token_embeddings = F.normalize(
+            self.contrastive_alignment_text_projection(
+                model_output.transformer_output.text_memory
+            ).transpose(0, 1),
+            p=2,
+            dim=-1,
+        )
+        contrastive_outputs = ContrastiveEmbeddingsOutput(
+            contrastive_query_embeddings, contrastive_token_embeddings
+        )
+
+        # Apply VQA heads to get answer predictions
+        answer_preds = self.vqa_heads(model_output.extra_embeddings.transpose(0, 1))
+
+        return MDETRVQAOutput(model_output, answer_preds, contrastive_outputs)
+
+
+def mdetr_for_vqa(
+    num_queries: int = 100,
+    num_classes: int = 255,
+    embedding_dim: int = 768,
+    transformer_d_model: int = 256,
+    transformer_num_heads: int = 8,
+    transformer_encoder_layers: int = 6,
+    transformer_decoder_layers: int = 6,
+    transformer_dim_feedforward: int = 2048,
+    transformer_dropout: float = 0.1,
+    return_intermediate_dec: bool = True,
+    contrastive_dim: int = 64,
+) -> MDETRForVQA:
+    hidden_dim = transformer_d_model
+    vqa_heads = mdetr_gqa_heads()
+    num_heads = len(vqa_heads.heads.keys())
+
+    model = mdetr_resnet101(
+        num_queries,
+        num_classes,
+        embedding_dim,
+        transformer_d_model,
+        transformer_num_heads,
+        transformer_encoder_layers,
+        transformer_decoder_layers,
+        transformer_dim_feedforward,
+        transformer_dropout,
+        return_intermediate_dec,
+        num_extra_query_embeddings=num_heads,
+    )
+    contrastive_alignment_image_projection = nn.Linear(hidden_dim, contrastive_dim)
+    contrastive_alignment_text_projection = nn.Linear(hidden_dim, contrastive_dim)
+
+    return MDETRForVQA(
+        model,
+        vqa_heads,
+        contrastive_alignment_image_projection,
+        contrastive_alignment_text_projection,
+    )
+
+
+class MDETRPhraseGroundingOutput(NamedTuple):
+    model_output: MDETRModelOutput
+    contrastive_embeddings: ContrastiveEmbeddingsOutput
+
+
+class MDETRForPhraseGrounding(nn.Module):
+    def __init__(
+        self,
+        model: MDETR,
+        contrastive_alignment_image_projection: nn.Module,
+        contrastive_alignment_text_projection: nn.Module,
+    ):
+        super().__init__()
+        self.model = model
+        self.contrastive_alignment_image_projection = (
+            contrastive_alignment_image_projection
+        )
+        self.contrastive_alignment_text_projection = (
+            contrastive_alignment_text_projection
+        )
+
+    def forward(
+        self,
+        images: List[Tensor],
+        text: List[Tensor],
+    ) -> MDETRPhraseGroundingOutput:
+
+        model_output = self.model(images, text)
+        final_hidden_state = model_output.transformer_output.decoder_hidden_states[-1]
+
+        contrastive_query_embeddings = F.normalize(
+            self.contrastive_alignment_image_projection(final_hidden_state),
+            p=2,
+            dim=-1,
+        )
+        contrastive_token_embeddings = F.normalize(
+            self.contrastive_alignment_text_projection(
+                model_output.transformer_output.text_memory
+            ).transpose(0, 1),
+            p=2,
+            dim=-1,
+        )
+        contrastive_outputs = ContrastiveEmbeddingsOutput(
+            contrastive_query_embeddings, contrastive_token_embeddings
+        )
+
+        return MDETRPhraseGroundingOutput(model_output, contrastive_outputs)
+
+
+def mdetr_for_phrase_grounding(
+    num_queries: int = 100,
+    num_classes: int = 255,
+    embedding_dim: int = 768,
+    transformer_d_model: int = 256,
+    transformer_num_heads: int = 8,
+    transformer_encoder_layers: int = 6,
+    transformer_decoder_layers: int = 6,
+    transformer_dim_feedforward: int = 2048,
+    transformer_dropout: float = 0.1,
+    return_intermediate_dec: bool = True,
+    contrastive_dim: int = 64,
+) -> MDETRForPhraseGrounding:
+    model = mdetr_resnet101(
+        num_queries,
+        num_classes,
+        embedding_dim,
+        transformer_d_model,
+        transformer_num_heads,
+        transformer_encoder_layers,
+        transformer_decoder_layers,
+        transformer_dim_feedforward,
+        transformer_dropout,
+        return_intermediate_dec,
+    )
+    hidden_dim = transformer_d_model
+    contrastive_alignment_image_projection = nn.Linear(hidden_dim, contrastive_dim)
+    contrastive_alignment_text_projection = nn.Linear(hidden_dim, contrastive_dim)
+
+    return MDETRForPhraseGrounding(
+        model,
+        contrastive_alignment_image_projection,
+        contrastive_alignment_text_projection,
+    )
